completed controller test, improved current_user validation in ShowContainer and memes_container_spec. Andrew & Leona

+++ app/javascript/react/components/ShowContainer.js
@@ -11,12 +11,19 @@ class ShowContainer extends Component {
       meme_is_current_user: false
     };
     this.deleteMeme = this.deleteMeme.bind(this)
-
   }
+
   componentDidMount() {
     let memeId = this.props.params.id;
 
-    fetch(`/api/v1/memes/${memeId}`)
+    fetch(`/api/v1/memes/${memeId}`, {
+      credentials: 'same-origin',
+      method: 'GET',
+      headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json'
+      }
+    })
       .then(response => {
         if (response.ok) {
           return response;
@@ -42,29 +49,34 @@ class ShowContainer extends Component {
     let memeToDelete = this.state.meme
     let memeId = this.state.meme.id;
 
-    fetch(`/api/v1/memes/${memeId}`, {
-      credentials: 'same-origin',
-      method: 'DELETE',
-      body: JSON.stringify(memeToDelete)
-    })
-      .then(response => {
-        if (response.ok) {
-          return response;
-        } else {
-          let errorMessage = `${response.status}(${response.statusText})` ,
-          error = new Error(errorMessage);
-          throw(error);
-        }
+    if (this.state.meme_is_current_user === true){
+      fetch(`/api/v1/memes/${memeId}`, {
+        credentials: 'same-origin',
+        method: 'DELETE',
+        headers: {
+          'Accept': 'application/json',
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify(memeToDelete)
       })
-      .then(response => response.json())
-      .then(body => {
-        return window.location.href = "/memes"
-      })
-      .catch(error => console.error(`Error in fetch: ${error.message}`))
+        .then(response => {
+          if (response.ok) {
+            return response;
+          } else {
+            let errorMessage = `${response.status}(${response.statusText})` ,
+            error = new Error(errorMessage);
+            throw(error);
+          }
+        })
+        .then(response => response.json())
+        .then(body => {
+            return window.location.href = "/memes"
+        })
+        .catch(error => console.error(`Error in fetch: ${error.message}`))
+    }
   }
 
   render() {
-
     return (
       <div>
         <h2>{this.state.meme.title}</h2>
